Adding error handling to UI
If an invalid token comes in we notify the user that the state is
invalid and ask them to try again. There is no UI button because we
don't know where they came from, and using the browser history to go
back will likely leave them in a broken state in the flow.

If they have been redirected by the rule back to this point, then they
likely tried to link accounts and accidentally chose an account with a
mismatched email (like a second gmail account or a friend's facebook).