Improving the specs and allowing handlers to be injected
I couldn't mock things out in a sane way so instead of loading the handlers
directly from the file in the server's initialization, I opted to import the
file as it was before, but pass the handlers to the server as an an argument
default if new handlers weren't explicitly passed.

Yes, this was indeed a situation of changing the code to appease the tests, but
I don't feel there is a degredation in readability. Full speed ahead!